refactor(domain): GameState 와 Decision 패키지 변경
