+++ src/main/kotlin/domain/Decision.kt
@@ -1,4 +1,4 @@
-package domain.constant
+package domain
 
 enum class Decision(val text: String) {
     YES("y"), NO("n")
