+++ src/test/kotlin/domain/constant/DecisionTest.kt
@@ -1,5 +1,6 @@
 package domain.constant
 
+import domain.Decision
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
